[Bug Fix] Kernel args alignment information was not taken into account when launching kernels. This fix manually parses through the code object and retrieves the kernel args alignment info before passing kernel args for a kernel launch.
Implementation is inspired from the OpenCL Runtime team
https://github.com/RadeonOpenCompute/ROCm-OpenCL-Runtime

Code object parsing now depends only on comgr and libelf-dev.
The comgr library is positioned to be shipped along with the
rocm-dev package, which ATMI depends on anyway. Moreover, this
approach gives us ample scope to support code object v3 as well
as v2 because comgr handles both.


Former-commit-id: bf53de1e4d0f3e42ae14df5f9ddf2740be9f460d